feat: parse from an arrray of points or a xy object
